+++ core/class/stickde3.class.php
@@ -37,6 +37,18 @@ class stickde3 extends eqLogic {
   public function postSave() {
   }
 
+  public function postUpdate() {
+      $stickde3Cmd = $this->getCmd(null, 'last');
+  if (!is_object($stickde3Cmd)) {
+			$stickde3Cmd = new stickde3Cmd();
+		}
+		$stickde3Cmd->setName(__('Dernière commande', __FILE__));
+		$stickde3Cmd->setLogicalId('last');
+		$stickde3Cmd->setEqLogic_id($this->getId());
+		$stickde3Cmd->setType('info');
+		$stickde3Cmd->setSubType('string');
+		$stickde3Cmd->save();
+  }
     public function callstickde3($_url) {
     $sock = socket_create(AF_INET, SOCK_DGRAM, SOL_UDP);
       log::add('stickde3', 'debug', 'Commande avant conversion:' . $_url);  
@@ -50,17 +62,16 @@ class stickde3 extends eqLogic {
     return $result;
   }
   
+
 }
 
 
 class stickde3Cmd extends cmd {
   public function execute($_options = array()) {
-    switch ($this->getType()) {
-      case 'action' :
-        $eqLogic = $this->getEqLogic();
+      $eqLogic = $this->getEqLogic();
+
         $eqLogic->callstickde3($this->getConfiguration('commande'));
-        
-    }
+        $eqLogic->checkAndUpdateCmd('last', $this->getname());
   }
 }
   ?>