+++ desktop/js/stickde3.js
@@ -82,7 +82,7 @@ $(".li_eqLogic").on('click', function (event) {
   }
   return false;
 });
-
+/*
 $("#bt_addstickde3Info").on('click', function(event) {
     var _cmd = {type: 'info'};
     addCmdToTable(_cmd);
@@ -94,7 +94,7 @@ $("#bt_addstickde3Action").on('click', function(event) {
 });
 
 
-/*
+
 $('#table_cmd tbody').delegate('tr .remove', 'click', function(event) {
     $(this).closest('tr').remove();
 });*/
@@ -135,36 +135,33 @@ function addCmdToTable(_cmd) {
   tr += '<span class="cmdAttr input-group-addon roundedRight" data-l1key="display" data-l2key="icon" style="font-size:19px;padding:0 5px 0 0!important;"></span>'
   tr += '</div>'
   tr += '</td>'
-
-
   tr += '<td class="type" type="' + init(_cmd.type) + '">' + jeedom.cmd.availableType();
   tr += '<span class="subType" subType="' + init(_cmd.subType) + '"></span></td>';
   tr += '</td>'
-  tr += '<td>';
-  tr += '<input class="cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="commande" style="width: 90%;display: inherit" ></input>';
+    tr += '<td>';
+    if (!isset(_cmd.type) || _cmd.type == 'action') {
+        tr += '<input class="cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="commande" style="width: 90%;display: inherit" ></input>';
+    }
   tr += '</td>';
   tr += '<td>'
-  tr += '<label class="checkbox-inline"><input type="checkbox" class="cmdAttr" data-l1key="isVisible" checked/>{{Afficher}}</label> '
+    tr += '<label class="checkbox-inline"><input type="checkbox" class="cmdAttr" data-l1key="isVisible" checked/>{{Afficher}}</label> '
+    if (!isset(_cmd.type) || _cmd.type == 'info') {
+        tr += '<span><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isHistorized" checked/>{{Historiser}}</label></span> ';
+    }
   tr += '</td>'
   tr += '<td>'
+  
   if (is_numeric(_cmd.id)) {
     tr += '<a class="btn btn-default btn-xs cmdAction" data-action="configure"><i class="fas fa-cogs"></i></a> '
     tr += '<a class="btn btn-default btn-xs cmdAction" data-action="test"><i class="fas fa-rss"></i> Tester</a>'
   }
   tr += '<i class="fas fa-minus-circle pull-right cmdAction cursor" data-action="remove" title="{{Supprimer la commande}}"></i></td>'
-  tr += '</tr>'
-  $('#table_cmd tbody').append(tr)
-  var tr = $('#table_cmd tbody tr').last()
-  jeedom.eqLogic.builSelectCmd({
-    id:  $('.eqLogicAttr[data-l1key=id]').value(),
-    filter: {type: 'info'},
-    error: function (error) {
-      $('#div_alert').showAlert({message: error.message, level: 'danger'})
-    },
-    success: function (result) {
-      tr.find('.cmdAttr[data-l1key=value]').append(result)
-      tr.setValues(_cmd, '.cmdAttr')
-      jeedom.cmd.changeType(tr, init(_cmd.subType))
+    tr += '</tr>';
+    $('#table_cmd tbody').append(tr);
+    $('#table_cmd tbody tr').last().setValues(_cmd, '.cmdAttr');
+    if (isset(_cmd.type)) {
+        $('#table_cmd tbody tr:last .cmdAttr[data-l1key=type]').value(init(_cmd.type));
     }
-  })
+    jeedom.cmd.changeType($('#table_cmd tbody tr').last(), init(_cmd.subType));
 }
+