+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 03/07/2022
+
+- Rajout de la commande info affichant la dernière commande réalisée depuis jeedom sur le stick
+
 # 15/05/2022
 
 - Création du plugin